Minor test refactor
